Change int byte

+++ dhis-2/dhis-support/dhis-support-db-migration/src/main/resources/org/hisp/dhis/db/migration/2.37/V2_37_36__Add_userinfo_dataviewmaxorganisationunitlevel.sql
@@ -1,2 +1,2 @@
 
-alter table "userinfo" add column if not exists "dataviewmaxorgunitlevel" int8 null;
+alter table "userinfo" add column if not exists "dataviewmaxorgunitlevel" int4 null;
